PG-2010 Replace project with a stub
We plan to, at least temporarily, retire this project due to lack of
faith in the quality of the the data it gathers, and to do that safely
without breaking people's installations we need to replace it with a
dummy module which does as little as possible.

The reason for this is because this extension is loaded via
shared_preload_libraries and if you just remove the extension from
people's computers when they upgrade Percona's PostgreSQL distribution
then PostgreSQL will fail to start after upgrading. So we will still
need to ship a shared library and to minimize maintenance we make it do
as little as possible.

This commit removes all logic expect the two SQL facing functions
exposed in the extension.